+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -1,18 +1,30 @@
 package com.techcourse.dao;
 
 import com.techcourse.domain.User;
+import jakarta.annotation.Nonnull;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.List;
 import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
 
 public class UserDao {
 
-    private static final Logger log = LoggerFactory.getLogger(UserDao.class);
+    private final RowMapper<User> userRowMapper = new RowMapper<>() {
+        @Nonnull
+        @Override
+        public User mapRow(final ResultSet resultSet, final int rowNumber) throws SQLException {
+            return new User(
+                resultSet.getLong("id"),
+                resultSet.getString("account"),
+                resultSet.getString("password"),
+                resultSet.getString("email")
+            );
+        }
+    };
 
     private final JdbcTemplate jdbcTemplate;
-    private final UserRowMapper userRowMapper = new UserRowMapper();
 
     public UserDao(final DataSource dataSource) {
         this.jdbcTemplate = new JdbcTemplate(dataSource);
