+++ jdbc/src/main/java/org/springframework/jdbc/core/RowMapper.java
@@ -2,12 +2,13 @@ package org.springframework.jdbc.core;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 @FunctionalInterface
 public interface RowMapper<T> {
 
-    @Nullable
+    @Nonnull
     T mapRow(final ResultSet resultSet, final int rowNumber) throws SQLException;
 }
